[Refactor] : repository 파일 위치 변경 (#5)
