+++ src/main/java/com/potato/balbambalbam/weaksoundtest/controller/WeakSoundTestController.java
@@ -3,7 +3,7 @@ package com.potato.balbambalbam.weaksoundtest.controller;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.potato.balbambalbam.weaksoundtest.dto.WeakSoundTestDto;
-import com.potato.balbambalbam.weaksoundtest.repository.WeakSoundTestRepository;
+import com.potato.balbambalbam.repository.WeakSoundTestRepository;
 import com.potato.balbambalbam.weaksoundtest.service.WeakSoundTestService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
