+++ src/main/java/com/potato/balbambalbam/weaksoundtest/service/WeakSoundTestService.java
@@ -2,7 +2,7 @@ package com.potato.balbambalbam.weaksoundtest.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.potato.balbambalbam.weaksoundtest.repository.WeakSoundTestRepository;
+import com.potato.balbambalbam.repository.WeakSoundTestRepository;
 import com.potato.balbambalbam.weaksoundtest.dto.WeakSoundTestDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
